Move function attributes to separate package
